requirements/development.txt
-r base.txt
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
responses==0.25.7
django-debug-toolbar==4.3.0
